relocate adminpage queries
