+++ src/Services/RestService.js
@@ -7,153 +7,6 @@ const instance = axios.create({
 });
 
 const RestService = (() => {
-    const createUserPrincipal = async (adminId, fname, mname, lname, username, email, password) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_USER}/create/principal`, {
-                adminId,
-                fname,
-                mname,
-                lname,
-                username,
-                email,
-                password,
-                position: "Principal"
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            })
-
-            if (response) {
-                console.log(response.data)
-            }
-
-            return response.status === 201;
-        } catch (error) {
-            console.error('Error creating user:', error);
-            if (error.response && error.response.status === 409) {
-                throw new Error("User with the same email or username already exists.");
-            } else {
-                throw new Error("Registration failed. Please try again later.");
-            }
-        }
-    };
-
-    const getSchoolName = async (name) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_SCHOOL}/name`, {
-                name
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error retrieving school:', error);
-            return null;
-        }
-    };
-
-    const getSchoolFullName = async (fullName) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_SCHOOL}/fullname`, {
-                fullName
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error retrieving school:', error);
-            return null;
-        }
-    };
-
-    const createSchool = async (name, fullName) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_SCHOOL}/create`, {
-                name,
-                fullName
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error creating school:', error);
-            return null;
-        }
-    };
-
-    const getPrincipal = async (schoolId) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_SCHOOL}/principal`, {
-                schoolId
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error creating school:', error);
-            return null;
-        }
-    };
-
-    const getUserByEmailUsername = async (email) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_USER}/schools`, {
-                emailOrUsername: email
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error creating school:', error);
-            return null;
-        }
-    };
-
-    const insertUserAssociation = async (userId, schoolId) => {
-        try {
-            const response = await instance.post(`${process.env.REACT_APP_API_URL_ASSOC}/insert`, {
-                userId,
-                schoolId
-            }, {
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            return response?.data || null;
-        } catch (error) {
-            console.error('Error creating school:', error);
-            return null;
-        }
-    };
-
-    const getSchools = async () => {
-        try {
-            const response = await instance.get(`${process.env.REACT_APP_API_URL_SCHOOL}/all`); // Adjust endpoint as needed
-            console.log('Fetched schools:', response.data);
-            return response.data;
-        } catch (error) {
-            console.error('Error fetching schools:', error);
-            throw new Error('Failed to fetch schools. Please try again later.');
-        }
-    };
-
     const createNotification = async (userId, message, type) => {
         try {
             const response = await instance.post(`${process.env.REACT_APP_API_URL_NOTIFICATION}/create`, {
@@ -212,14 +65,6 @@ const RestService = (() => {
 
 
     return {
-        createUserPrincipal,
-        getSchoolName,
-        getSchoolFullName,
-        createSchool,
-        getPrincipal,
-        getUserByEmailUsername,
-        insertUserAssociation,
-        getSchools,
         createNotification,
         getNotificationsForUser,
         markNotificationAsRead,
